+++ sdk/resourcemanager/compute/armcompute/zz_generated_models_serde.go
@@ -373,16 +373,26 @@ func (c CloudServiceVaultSecretGroup) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
+// MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageList.
+func (c CommunityGalleryImageList) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "nextLink", c.NextLink)
+	populate(objectMap, "value", c.Value)
+	return json.Marshal(objectMap)
+}
+
 // MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageProperties.
 func (c CommunityGalleryImageProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "disallowed", c.Disallowed)
 	populateTimeRFC3339(objectMap, "endOfLifeDate", c.EndOfLifeDate)
+	populate(objectMap, "eula", c.Eula)
 	populate(objectMap, "features", c.Features)
 	populate(objectMap, "hyperVGeneration", c.HyperVGeneration)
 	populate(objectMap, "identifier", c.Identifier)
 	populate(objectMap, "osState", c.OSState)
 	populate(objectMap, "osType", c.OSType)
+	populate(objectMap, "privacyStatementUri", c.PrivacyStatementURI)
 	populate(objectMap, "purchasePlan", c.PurchasePlan)
 	populate(objectMap, "recommended", c.Recommended)
 	return json.Marshal(objectMap)
@@ -403,6 +413,9 @@ func (c *CommunityGalleryImageProperties) UnmarshalJSON(data []byte) error {
 		case "endOfLifeDate":
 			err = unpopulateTimeRFC3339(val, &c.EndOfLifeDate)
 			delete(rawMsg, key)
+		case "eula":
+			err = unpopulate(val, &c.Eula)
+			delete(rawMsg, key)
 		case "features":
 			err = unpopulate(val, &c.Features)
 			delete(rawMsg, key)
@@ -418,6 +431,9 @@ func (c *CommunityGalleryImageProperties) UnmarshalJSON(data []byte) error {
 		case "osType":
 			err = unpopulate(val, &c.OSType)
 			delete(rawMsg, key)
+		case "privacyStatementUri":
+			err = unpopulate(val, &c.PrivacyStatementURI)
+			delete(rawMsg, key)
 		case "purchasePlan":
 			err = unpopulate(val, &c.PurchasePlan)
 			delete(rawMsg, key)
@@ -432,11 +448,21 @@ func (c *CommunityGalleryImageProperties) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageVersionList.
+func (c CommunityGalleryImageVersionList) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "nextLink", c.NextLink)
+	populate(objectMap, "value", c.Value)
+	return json.Marshal(objectMap)
+}
+
 // MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageVersionProperties.
 func (c CommunityGalleryImageVersionProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populateTimeRFC3339(objectMap, "endOfLifeDate", c.EndOfLifeDate)
+	populate(objectMap, "excludeFromLatest", c.ExcludeFromLatest)
 	populateTimeRFC3339(objectMap, "publishedDate", c.PublishedDate)
+	populate(objectMap, "storageProfile", c.StorageProfile)
 	return json.Marshal(objectMap)
 }
 
@@ -452,8 +478,14 @@ func (c *CommunityGalleryImageVersionProperties) UnmarshalJSON(data []byte) erro
 		case "endOfLifeDate":
 			err = unpopulateTimeRFC3339(val, &c.EndOfLifeDate)
 			delete(rawMsg, key)
+		case "excludeFromLatest":
+			err = unpopulate(val, &c.ExcludeFromLatest)
+			delete(rawMsg, key)
 		case "publishedDate":
 			err = unpopulateTimeRFC3339(val, &c.PublishedDate)
+			delete(rawMsg, key)
+		case "storageProfile":
+			err = unpopulate(val, &c.StorageProfile)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -461,6 +493,14 @@ func (c *CommunityGalleryImageVersionProperties) UnmarshalJSON(data []byte) erro
 		}
 	}
 	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type CommunityGalleryImageVersionStorageProfile.
+func (c CommunityGalleryImageVersionStorageProfile) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "dataDiskImages", c.DataDiskImages)
+	populate(objectMap, "osDiskImage", c.OSDiskImage)
+	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type CommunityGalleryInfo.